requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.4.0
Flask==2.2.2
Flask-Cors==4.0.1
Flask-Session==0.4.0
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
pip==24.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.19
waitress==3.0.0
Werkzeug==2.2.2
yarl==1.9.4
pyjwt==2.8.0